+++ Hazard-Renderer/src/Backend/OpenGL/OpenGLFramebuffer.h
@@ -18,11 +18,11 @@ namespace HazardRenderer::OpenGL
 		virtual void Unbind() override;
 
 		virtual void Resize(uint32_t width, uint32_t height, bool force = false) override;
-		virtual void BindTexture(uint32_t attachmentIndex = 0, uint32_t slot = 0) const override {};
+		virtual void BindTexture(uint32_t attachmentIndex = 0, size_t slot = 0) const override {};
 
 		virtual uint32_t GetWidth() const override { return m_Specs.Width; };
 		virtual uint32_t GetHeight() const override { return m_Specs.Height; };
-		uint32_t GetColorAttachmentCount() override { return m_ColorAttachments.size(); }
+		size_t GetColorAttachmentCount() override { return m_ColorAttachments.size(); }
 
 		virtual Ref<Image2D> GetImage(uint32_t index = 0) const override { return m_ColorImages[index].As<Image2D>(); };
 		virtual FrameBufferSpecification& GetSpecification() override { return m_Specs; }
