+++ Hazard/src/Hazard/Rendering/LineRenderer.cpp
@@ -37,8 +37,8 @@ namespace Hazard
 		HZR_PROFILE_FUNCTION();
 		if (!m_LineBatch.GetCount()) return;
 
-		uint32_t size = m_LineBatch.GetDataSize();
-		uint32_t elements = size / sizeof(LineVertex);
+		uint64_t size = m_LineBatch.GetDataSize();
+		uint64_t elements = size / sizeof(LineVertex);
 		m_VertexBuffer->SetData(m_LineBatch.GetData(), size);
 
 		HRenderer::SubmitMesh(glm::mat4(1.0f), m_VertexBuffer, m_Pipeline, m_LineBatch.GetCount());
